% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Inverse Value Matching}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector or NULL: values to be matched}

\item{y}{vector or NULL: values to be matched against}
}
\value{
Logical vector, indicating if value x was not found in y
}
\description{
Complement of \code{\%in\%}. Returns elements of \code{x} that are not in \code{y}.
}
\examples{
1 \%nin\% 2:6
1 \%nin\% 1:5
}
