% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-not-in-grapes.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not Value Matching}
\usage{
... \%!in\% NA
}
\arguments{
\item{x}{vector or NULL: values to be matched}

\item{table}{vector or NULL: values to be matched against}
}
\value{
Logical vector, indicating if value x was not found in table.
}
\description{
Check if the left hand side value is present in the right hand side. Return FALSE if it is, TRUE if not. Negation of \%in\%.
}
\examples{
1 \%!in\% 2:6
1 \%!in\% 1:5
}
