% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinDataDictionary.R
\name{joinDataWithDictionary}
\alias{joinDataWithDictionary}
\title{Combine dashboard data with dictionary}
\usage{
joinDataWithDictionary(consol_dat, data_dict)
}
\arguments{
\item{consol_dat}{Dashboard database}

\item{data_dict}{The data dictionary}
}
\value{
A data.frame
}
\description{
Combine dashboard data with dictionary
}
\examples{
\dontrun{
joinDataWithDictionary(consol_dat, data_dict)
}
}
