% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameProcessing.R
\name{removeCompoundSeparators}
\alias{removeCompoundSeparators}
\title{Remove Hyphens, Spaces, and Apostrophes from Compound Last Names}
\usage{
removeCompoundSeparators(df, last_name_column = "last_name")
}
\arguments{
\item{df}{A dataframe containing a last name column}

\item{last_name_column}{Character string specifying the name of the column
containing the last names. Default is "last_name".}
}
\value{
A dataframe with the compound last name separators removed
}
\description{
This function removes hyphens, spaces, and apostrophes from compound last names to create
a single continuous last name string. Useful for standardizing names like
"Van Der Berg" to "VanDerBerg", "Smith-Jones" to "SmithJones", or "O'Connor" to "OConnor".
}
\examples{
# After using splitNames function
df <- data.frame(
  first_name = c("JOHN", "MARY", "ROBERT"),
  last_name = c("VAN DER BERG", "SMITH-JONES", "O'CONNOR"),
  middle_name = c("", "ELIZABETH", ""),
  suffix = c("", "", "JR")
)
result <- removeCompoundSeparators(df)

# Custom column name
df <- data.frame(surname = c("DE LA CRUZ", "MARTINEZ-RODRIGUEZ"))
result <- removeCompoundSeparators(df, last_name_column = "surname")

}
