% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyDecimal.R
\name{specifyDecimal}
\alias{specifyDecimal}
\title{Specify number of decimal places to show}
\usage{
specifyDecimal(x, digits)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{The number of decimals to display}
}
\value{
A character vector
}
\description{
Specify number of decimal places to show
}
\examples{
specifyDecimal(pi, 1)
specifyDecimal(pi, 2)
specifyDecimal(pi, 3)
}
