% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_remake.R
\name{tar_remake}
\alias{tar_remake}
\title{Make and load a target}
\usage{
tar_remake(names)
}
\arguments{
\item{names}{The name(s) of the target to be remade. Accepts tidyselect helpers.}
}
\value{
If the target is successfully made, loads the target into the workspace.
}
\description{
Make and load a target from a targets workflow
}
