% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_remake.R
\name{tar_remake}
\alias{tar_remake}
\title{Make and load a target}
\usage{
tar_remake(names, notify = FALSE, reporter = "balanced")
}
\arguments{
\item{names}{The name(s) of the target to be remade. Accepts tidyselect helpers.}

\item{notify}{Pop up a windows notification that the pipeline is done. Only works on Windows}

\item{reporter}{Determinse how verbose the targets reporter is. Defaults to 'balanced'.}
}
\value{
If the target is successfully made, loads the target into the workspace.
}
\description{
Make and load a target from a targets workflow
}
\examples{
# Example stolen from the targets documentation
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
  # for CRAN
  tar_dir({
    # tar_dir() runs code from a temp dir for CRAN.
    tar_script({
      list(tar_target(y1, 1 + 1),
           tar_target(y2, 1 + 1),
           tar_target(z, y1 + y2))
    }, ask = FALSE)
    tar_remake(starts_with("y")) # Only processes y1 and y2.
  })
  }
\dontrun{
tar_remake(starts_with('model'))
}
}
